Reuse FrameUtils in GM algo
